♻️ Refactor: Confirm 컴포넌트 적용
